urlshortner/requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
decorator==5.0.9
Django==3.2.5
django-restframework==0.0.1
hashids==1.3.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
numpy==1.21.1
pandas==1.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
